refactor(be-auth): rename auth module to identity to avoid collision with django.contrib.auth
