+++ backend/rotini/identity/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class AuthConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "auth"
+    name = "identity"
