+++ backend/rotini/identity/jwt_test.py
@@ -2,17 +2,17 @@ import pytest
 import freezegun
 import jwt
 
-import auth.jwt
+import identity.jwt
 
 
 @freezegun.freeze_time("2012-01-01")
 def test_generates_and_decodes_token_token():
     MOCK_USER_ID = 1
-    token = auth.jwt.generate_token_for_user(MOCK_USER_ID)
+    token = identity.jwt.generate_token_for_user(MOCK_USER_ID)
 
     assert token is not None
 
-    token_data = auth.jwt.decode_token(token)
+    token_data = identity.jwt.decode_token(token)
 
     assert token_data["user_id"] == MOCK_USER_ID
 
@@ -21,9 +21,9 @@ def test_token_decode_fails_if_expired():
     MOCK_USER_ID = 1
 
     with freezegun.freeze_time("2012-01-01"):
-        token = auth.jwt.generate_token_for_user(MOCK_USER_ID)
+        token = identity.jwt.generate_token_for_user(MOCK_USER_ID)
 
     assert token is not None
 
     with pytest.raises(jwt.ExpiredSignatureError):
-        auth.jwt.decode_token(token)
+        identity.jwt.decode_token(token)
