+++ backend/rotini/identity/views.py
@@ -1,11 +1,12 @@
 import logging
+import uuid
 
 import django.http
 import django.contrib.auth
 import rest_framework.views
 import rest_framework.status
 
-import auth.jwt
+import identity.jwt
 
 AuthUser = django.contrib.auth.get_user_model()
 
@@ -45,7 +46,8 @@ class SessionListView(rest_framework.views.APIView):
         if user is not None:
             django.contrib.auth.login(request, user)
 
-            token = auth.jwt.generate_token_for_user(user_id=user.id)
+            token = identity.jwt.generate_token_for_user(user_id=user.id)
+
             response = django.http.HttpResponse(status=201)
 
             response.set_cookie(
